refactor: 함수명 변경
