+++ app/src/main/java/com/sopt/now/compose/SignUp/SignUpActivity.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.SignUp
 
 import android.content.Context
 import android.content.Intent
@@ -36,6 +36,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.sopt.now.compose.Login.LoginActivity
 import com.sopt.now.compose.ui.theme.NOWSOPTAndroidTheme
 
 class SignUpActivity : ComponentActivity() {
@@ -48,7 +49,7 @@ class SignUpActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    SignUpContent(viewModel)
+                    SignUpScreen(viewModel)
                 }
             }
         }
@@ -56,7 +57,7 @@ class SignUpActivity : ComponentActivity() {
 }
 
 @Composable
-fun SignUpContent(viewModel: SignUpViewModel) {
+fun SignUpScreen(viewModel: SignUpViewModel) {
     val signUpSuccess by viewModel.signUpSuccess.observeAsState(false)
     val context = LocalContext.current
     val id = remember { mutableStateOf("") }
@@ -184,12 +185,4 @@ fun moveToLoginActivity(
         putExtra("ADDRESS", address)
     }
     context.startActivity(intent)
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    NOWSOPTAndroidTheme {
-
-    }
 }
