+++ app/src/main/java/com/sopt/now/compose/SignUp/SignUpViewModel.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.SignUp
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
